double everywhere in ElementMonde

+++ src/main/java/fr/ul/acl/escape/monde/ElementMonde.java
@@ -2,38 +2,38 @@ package fr.ul.acl.escape.monde;
 
 public abstract class ElementMonde {
 
-    protected final int hauteur;
-    protected final int largeur;
-    protected float x, y;
+    protected final double hauteur;
+    protected final double largeur;
+    protected double x, y;
 
-    public ElementMonde(float x, float y, int hauteur, int largeur) {
+    public ElementMonde(double x, double y, double hauteur, double largeur) {
         this.x = x;
         this.y = y;
         this.hauteur = hauteur;
         this.largeur = largeur;
     }
 
-    public float getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(float x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public float getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(float y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public int getHauteur() {
+    public double getHauteur() {
         return hauteur;
     }
 
-    public int getLargeur() {
+    public double getLargeur() {
         return largeur;
     }
