+++ src/main/java/fr/ul/acl/escape/monde/Mur.java
@@ -1,7 +1,7 @@
 package fr.ul.acl.escape.monde;
 
 public class Mur extends Terrain {
-    public Mur(float x, float y, int hauteur, int largeur) {
+    public Mur(double x, double y, double hauteur, double largeur) {
         super(x, y, hauteur, largeur);
     }
