+++ src/main/java/fr/ul/acl/escape/monde/Personnage.java
@@ -3,9 +3,9 @@ package fr.ul.acl.escape.monde;
 import fr.ul.acl.escape.monde.exceptions.MouvementNullException;
 
 public abstract class Personnage extends ElementMonde {
-    protected float vitesse;
+    protected double vitesse;
 
-    public Personnage(float x, float y, int hauteur, int largeur, float vitesse) {
+    public Personnage(double x, double y, double hauteur, double largeur, double vitesse) {
         super(x, y, hauteur, largeur);
         vitesse = 1;
         this.vitesse = vitesse;
